+++ NEWS.md
@@ -1,5 +1,8 @@
 # nlmixr2est (development version)
 
+- A bug in model piping which did not allow models to be appended to was fixed
+  (rxode2#364)
+
 # nlmixr2est 2.1.3
 
 - Allows `$etaH` and related family to be integrated into a `saem` fit